Consolidate database entry point
