+++ webapp/src/routes/configure/-lib/login-rate-limit.server.test.ts
@@ -6,7 +6,7 @@ import * as Layer from "effect/Layer"
 import { RateLimiter } from "effect/unstable/persistence"
 import { SqlError, UnknownError } from "effect/unstable/sql/SqlError"
 
-import { type EffectDatabase, effectDatabase } from "@/db/effect.server"
+import { type EffectDatabase, effectDatabase } from "@/db"
 
 import { consumeOperatorLoginRateLimit } from "./login-rate-limit.server.ts"
